+++ CHANGELOG.md
@@ -5,3 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.0] - 2020-05-21
+### Added
+ - Composer setup
+ - Connection abstract wrapper class
+ - Migrator system with interface and abstract class
+
+[Unreleased]: https://github.com/AymDev/Fregata/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/AymDev/Fregata/releases/tag/v0.1.0